plant_detection/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
filelock==3.18.0
fsspec==2025.3.0
gunicorn==23.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2025.1
sendgrid==6.11.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
tzdata==2025.1
whitenoise==6.9.0
